feat: split mongodb service into new crate (RFC-6828)
